# Optimizes loadbalanced backend string